feat: add admin permission check on inter-canisters calls

+++ studio/canisters/slides/slides.mo
@@ -1,4 +1,5 @@
 import Iter "mo:base/Iter";
+import Principal "mo:base/Principal";
 
 import Error "mo:base/Error";
 
@@ -35,12 +36,20 @@ actor Slide {
     };
 
     public shared({ caller }) func del(slideId : SlideId) : async (Bool) {
-        let exists: Bool = await deleteSlide(caller, slideId);
+        let admin : Principal = Principal.fromActor(Slide);
+        
+        let exists: Bool = await delAdmin(admin, caller, slideId);
 
         return exists;
     };
 
-    public func deleteSlide(user: Principal, slideId : SlideId) : async (Bool) {
+    public func delAdmin(caller: Principal, user: Principal, slideId : SlideId) : async (Bool) {
+        let admin : Principal = Principal.fromActor(Slide);
+
+        if (caller != admin) {
+            throw Error.reject("This function requires administrative privileges.");
+        };
+
         let exists: Bool = await store.deleteSlide(user, slideId);
 
         return exists;
